fix: address round-4 review — approval count drift, lockout composable, 401 redirect
- Fix approval total.value++ outside filter guard (count drift bug)
- Extract useLoginLockout composable from Login/Setup pages (dedup + use HTTP status codes instead of string matching)
- Preserve redirect URL in 401 fallback handler
