+++ web/src/views/LoginPage.vue
@@ -1,39 +1,22 @@
 <script setup lang="ts">
-import { ref, computed, onUnmounted } from 'vue'
+import { ref } from 'vue'
 import { useRouter } from 'vue-router'
 import InputText from 'primevue/inputtext'
 import Button from 'primevue/button'
 import { useToast } from 'primevue/usetoast'
 import { useAuthStore } from '@/stores/auth'
 import { getErrorMessage } from '@/utils/errors'
-import { MIN_PASSWORD_LENGTH, LOGIN_MAX_ATTEMPTS, LOGIN_LOCKOUT_MS } from '@/utils/constants'
+import { MIN_PASSWORD_LENGTH } from '@/utils/constants'
+import { useLoginLockout } from '@/composables/useLoginLockout'
 
 const router = useRouter()
 const auth = useAuthStore()
 const toast = useToast()
+const { locked, checkAndClearLockout, recordFailure, reset } = useLoginLockout()
 
 const username = ref('')
 const password = ref('')
 const error = ref<string | null>(null)
-// Client-side lockout is a UX hint only — it resets on page refresh.
-// Real brute-force protection is enforced server-side via rate limiting.
-const attempts = ref(0)
-const lockedUntil = ref<number | null>(null)
-
-// Reactive clock so `locked` re-evaluates when lockout expires
-const now = ref(Date.now())
-const clockTimer = setInterval(() => { now.value = Date.now() }, 1000)
-onUnmounted(() => clearInterval(clockTimer))
-
-const locked = computed(() => !!(lockedUntil.value && now.value < lockedUntil.value))
-
-function checkAndClearLockout(): boolean {
-  if (lockedUntil.value && Date.now() >= lockedUntil.value) {
-    lockedUntil.value = null
-    attempts.value = 0
-  }
-  return locked.value
-}
 
 async function handleLogin() {
   if (checkAndClearLockout()) {
@@ -43,27 +26,18 @@ async function handleLogin() {
   error.value = null
   try {
     await auth.login(username.value, password.value)
-    attempts.value = 0
+    reset()
     if (auth.mustChangePassword) {
       toast.add({ severity: 'warn', summary: 'Password change required', life: 5000 })
     }
     router.push('/')
   } catch (err) {
-    const msg = getErrorMessage(err)
-    // Only count credential failures (4xx) toward lockout, not network/5xx errors
-    const isCredentialError = msg !== 'Network error. Please check your connection.' &&
-      msg !== 'A server error occurred. Please try again later.' &&
-      msg !== 'Service temporarily unavailable. Please try again later.'
-    if (isCredentialError) {
-      attempts.value++
-      if (attempts.value >= LOGIN_MAX_ATTEMPTS) {
-        lockedUntil.value = Date.now() + LOGIN_LOCKOUT_MS
-        attempts.value = 0
-        error.value = `Too many failed attempts. Please wait ${LOGIN_LOCKOUT_MS / 1000} seconds.`
-        return
-      }
+    const lockoutMsg = recordFailure(err)
+    if (lockoutMsg) {
+      error.value = lockoutMsg
+      return
     }
-    error.value = msg
+    error.value = getErrorMessage(err)
   }
 }
